# sqlc-grpc-lern/api/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "my",
    "description": "Boilerplate code generated by **sqlc-grpc**. Modify _proto/*.proto_ files then run `buf generate` to change the services interface.",
    "version": "1.0",
    "contact": {
      "name": "sqlc-grpc",
      "url": "https://github.com/walterwanderley/sqlc-grpc"
    }
  },
  "tags": [
    {
      "name": "AuthorService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/author": {
      "post": {
        "operationId": "AuthorService_CreateAuthor",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateAuthorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateAuthorRequest"
            }
          }
        ],
        "tags": [
          "AuthorService"
        ]
      }
    },
    "/author/{id}": {
      "get": {
        "operationId": "AuthorService_GetAuthor",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1Author"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "AuthorService"
        ]
      },
      "delete": {
        "operationId": "AuthorService_DeleteAuthor",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteAuthorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "AuthorService"
        ]
      }
    },
    "/authors": {
      "get": {
        "operationId": "AuthorService_ListAuthors",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/v1Author"
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "AuthorService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Author": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "companyId": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "string",
          "format": "int64"
        },
        "emptyCol": {
          "type": "integer",
          "format": "int32"
        },
        "defaultCol": {
          "type": "integer",
          "format": "int32"
        },
        "size1": {
          "type": "integer",
          "format": "int32"
        },
        "defaultCol1": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "type1": {
          "type": "string"
        }
      }
    },
    "v1CreateAuthorRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "size": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1CreateAuthorResponse": {
      "type": "object"
    },
    "v1DeleteAuthorResponse": {
      "type": "object"
    },
    "v1GetAuthorResponse": {
      "type": "object",
      "properties": {
        "author": {
          "$ref": "#/definitions/v1Author"
        }
      }
    },
    "v1ListAuthorsResponse": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Author"
          }
        }
      }
    }
  }
}